+++ packages/microsoft-security-adapter-e2e/e2e_test/adapter.test.ts
@@ -27,8 +27,8 @@ import {
   microsoftSecurityCleanupChangeErrorFilter,
   microsoftSecurityCleanupValidationFilter,
   microsoftSecurityDeployChangeErrorFilter,
+  verifyInstanceValues,
 } from './helpers'
-import { verifyInstanceValues } from './verificationUtils'
 import { modificationChangesBeforeAndAfterOverrides } from './mock_elements'
 
 const log = logger(module)
