+++ packages/microsoft-security-adapter/jest.config.js
@@ -14,7 +14,7 @@ module.exports = deepMerge(require('../../jest.base.config.js'), {
   testEnvironment: undefined,
   coverageThreshold: {
     global: {
-      statements:  97.14,
+      statements: 97.14,
       branches: 88.06,
       functions: 91.64,
       lines: 97.03,
